requirement.txt
pyserial
mysql-connector 
Flask
